# Make ThriftContextFactory more efficient